cleanup: break up annotations into files
- rename Annotations.drawAll -> Annotations.draw
- make old Annotatios.draw a private method in annotations/draw.js
